Twitter/twitter-clone-backend-django/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-environ>=0.11
django-filter>=24.2
drf-spectacular>=0.27
psycopg[binary]>=3.1
gunicorn>=22.0
kafka-python>=2.0
redis>=5.0
requests>=2.32
django-prometheus>=2.3
channels>=4.1
channels-redis>=4.2
uvicorn>=0.30
websockets>=12.0
